Getting latest dev file for testing

+++ src/Server/Server.java
@@ -1,8 +1,10 @@
 package Server;
 
 import ControlPanel.User;
+import ControlPanel.schedule.CalanderScheduleGUI;
 import Server.Reply.*;
 import Server.Request.*;
+
 import java.io.*;
 import java.net.ServerSocket;
 import java.net.Socket;
@@ -16,6 +18,7 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Properties;
 import java.util.Random;
+
 import static Server.UserSQL.*;
 
 public class Server {
@@ -130,15 +133,13 @@ public class Server {
             statement.execute(CREATE_USER_TABLE);
             statement.execute(CREATE_SCHEDULE_TABLE);
 
-//################code below is just to create a test user with no name or password for testing
-
             // Username and Password are added.
             try {
                 ResultSet resultSet = statement.executeQuery("SELECT * FROM user");
                 String testAdmin = "admin";
                 String testPassword = "test1";
                 String testSaltString = randomString();
-                Boolean AdminExists = false;
+                boolean AdminExists = false;
                 String hashedPassword = hashAString(testPassword + testSaltString);
 
                 while (resultSet.next()) {
@@ -165,7 +166,6 @@ public class Server {
             } catch (NoSuchAlgorithmException e) {
                 e.printStackTrace();
             }
-//#########################above code is just to create a test user with no name or password for testing
         } catch (SQLException ex) {
             ex.printStackTrace();
         }
@@ -176,7 +176,6 @@ public class Server {
      *
      * @param hashString salted password to be hashed
      * @return salted hashed password
-     * @author Nic
      * @throws NoSuchAlgorithmException
      * @author Nic
      */
@@ -473,7 +472,6 @@ public class Server {
 
         }
         // If the request is an instance of list all users request
-
         else if (clientRequest instanceof ListUserRequest){
             ListUserRequest listUserRequest = (ListUserRequest) clientRequest;
             SessionToken sessionToken = findSessionToken(listUserRequest.getSessionToken());
@@ -487,16 +485,13 @@ public class Server {
             else{
                 // Reset the used time of the session token.
                 resetSessionTokenTime(sessionToken);
-
                 for (int i = 0; i <= sessionTokens.size(); i++) {
                     if (sessionTokens.get(i).getSessionTokenString().equals(sessionToken.getSessionTokenString())) {
                         sessionTokens.get(i).setUsedTime(LocalDateTime.now());
                         break;
                     }
                 }
-
                 ListUserReply listUserReply = new ListUserReply(sessionToken, listUserSQL(), true);
-
                 oos.writeObject(listUserReply);
                 oos.flush();
             }
@@ -516,16 +511,21 @@ public class Server {
                 sessionToken.setUsedTime(LocalDateTime.now());
                 GeneralReply generalReply;
 
-                // Execute create/edit query and return general reply indicating success
-                try {
-                    BillboardSQL bb = new BillboardSQL();
-                    bb.CreateBillboard(temp.getBillboardName(), temp.getUserName(), temp.getTextColour(), temp.getBackgroundColour(),
-                            temp.getMessage(), temp.getImage(), temp.getInformation(), temp.getInformationColour());
-                    generalReply = new GeneralReply(sessionToken,true);
-                } catch (Exception e) {
+                // if the user has the permission to create a billboard
+                if (temp.getCreateBillboardPermission()) {
+                    // Execute create/edit query and return general reply indicating success
+                    try {
+                        BillboardSQL bb = new BillboardSQL();
+                        bb.CreateBillboard(temp.getBillboardName(), temp.getUserName(), temp.getTextColour(), temp.getBackgroundColour(),
+                                temp.getMessage(), temp.getImage(), temp.getInformation(), temp.getInformationColour());
+                        generalReply = new GeneralReply(sessionToken,true);
+                    } catch (Exception e) {
+                        generalReply = new GeneralReply(sessionToken,false);
+                    }
+                }
+                else {
                     generalReply = new GeneralReply(sessionToken,false);
                 }
-
                 oos.writeObject(generalReply);
             }
             oos.flush();
@@ -534,6 +534,7 @@ public class Server {
         else if (clientRequest instanceof EditBBRequest) {
             EditBBRequest temp = (EditBBRequest) clientRequest;
             SessionToken sessionToken = findSessionToken(temp.getSessionToken());
+            BillboardSQL bb = new BillboardSQL();
 
             // Find the session token in the list.
             for(int i  = 0; i <= sessionTokens.size(); i++){
@@ -556,20 +557,44 @@ public class Server {
                 try {
                     // update edited billboard
                     if (temp.getBillboardName() == null) {
-                        BillboardSQL bb = new BillboardSQL();
                         bb.EditBillboard(BillBoardName, temp.getEditTextColour(), temp.getEditBGColour(),
                                 temp.getEditMsg(), temp.getEditImg(), temp.getEditInfo(), temp.getEditInfoColour());
+                        new MakeXMLFile(BillBoardName);
                         generalReply = new GeneralReply(temp.getSessionToken(),true);
                         oos.writeObject(generalReply);
                     }
                     // return the contents of the billboard
                     else {
                         BillBoardName = temp.getBillboardName();
-                        BillboardSQL bb = new BillboardSQL();
-                        bb.EditBillboard(BillBoardName);
-                        EditBBReply editBBReply = new EditBBReply(temp.getSessionToken(), bb.textColour, bb.backgroundColour, bb.message,
-                                bb.image, bb.information, bb.informationColour);
-                        oos.writeObject(editBBReply);
+                        boolean checkOriginOwner = bb.checkBillboardUser(BillBoardName, temp.getLoginUser());
+                        // if the user is the owner of the billboard
+                        if (checkOriginOwner) {
+                            // if the user has edit all billboards or create billboard permission
+                            if (temp.getEditAllBillboardsPermission() || (temp.getCreateBillboardPermission() && !CalanderScheduleGUI.IsCurrentlyScheduled(BillBoardName))) {
+                                bb.EditBillboard(BillBoardName);
+                                EditBBReply editBBReply = new EditBBReply(temp.getSessionToken(), bb.textColour, bb.backgroundColour, bb.message,
+                                        bb.image, bb.information, bb.informationColour);
+                                oos.writeObject(editBBReply);
+                            }
+                            else {
+                                generalReply = new GeneralReply(temp.getSessionToken(),false);
+                                oos.writeObject(generalReply);
+                            }
+                        }
+                        // if the user is editing billboard owned by other user
+                        else {
+                            // if the user has edit all billboards permission
+                            if (temp.getEditAllBillboardsPermission()) {
+                                bb.EditBillboard(BillBoardName);
+                                EditBBReply editBBReply = new EditBBReply(temp.getSessionToken(), bb.textColour, bb.backgroundColour, bb.message,
+                                        bb.image, bb.information, bb.informationColour);
+                                oos.writeObject(editBBReply);
+                            }
+                            else {
+                                generalReply = new GeneralReply(temp.getSessionToken(),false);
+                                oos.writeObject(generalReply);
+                            }
+                        }
                     }
                 } catch (Exception e) {
                     generalReply = new GeneralReply(temp.getSessionToken(),false);
@@ -582,6 +607,7 @@ public class Server {
         else if (clientRequest instanceof DeleteBBRequest) {
             DeleteBBRequest temp = (DeleteBBRequest) clientRequest;
             SessionToken sessionToken = findSessionToken(temp.getSessionToken());
+            BillboardSQL bb = new BillboardSQL();
 
             // Remove session token from the list and send a logout request if it expired.
             if (!tokenCheck(temp.getSessionToken())) {
@@ -595,13 +621,39 @@ public class Server {
 
                 // Execute delete query and return general reply indicating success
                 try {
-                    BillboardSQL bb = new BillboardSQL();
-                    bb.DeleteBillboard(temp.getBillboardName());
-                    generalReply = new GeneralReply(sessionToken,true);
+                    BillBoardName = temp.getBillboardName();
+                    boolean checkOriginOwner = bb.checkBillboardUser(BillBoardName, temp.getLoginUser());
+                    // if the user is the owner of the billboard
+                    if (checkOriginOwner) {
+                        // if the user has edit all billboards or create billboard permission
+                        if (temp.getEditAllBillboardsPermission() || (temp.getCreateBillboardPermission() && !CalanderScheduleGUI.IsCurrentlyScheduled(BillBoardName))) {
+                            bb.DeleteBillboard(temp.getBillboardName());
+                            File deleteFile = new File("src/xmlBillboards/" + temp.getBillboardName() + ".xml");
+                            deleteFile.delete();
+                            generalReply = new GeneralReply(sessionToken,true);
+                        }
+                        else {
+                            generalReply = new GeneralReply(temp.getSessionToken(),false);
+                        }
+                    }
+                    // if the user is deleting billboard owned by other user
+                    else {
+                        // if the user has edit all billboards permission
+                        if (temp.getEditAllBillboardsPermission()) {
+                            bb.DeleteBillboard(temp.getBillboardName());
+                            File deleteFile = new File("src/xmlBillboards/" + temp.getBillboardName() + ".xml");
+                            deleteFile.delete();
+                            generalReply = new GeneralReply(temp.getSessionToken(),true);
+                        }
+                        else {
+                            generalReply = new GeneralReply(temp.getSessionToken(),false);
+                        }
+                    }
+                    oos.writeObject(generalReply);
                 } catch (Exception e) {
                     generalReply = new GeneralReply(sessionToken,false);
+                    oos.writeObject(generalReply);
                 }
-                oos.writeObject(generalReply);
             }
             oos.flush();
         }
@@ -652,7 +704,7 @@ public class Server {
                 // Return a reply object to the client containing JTable of billboards
                 try {
                     BillboardSQL bb = new BillboardSQL();
-                    ListBBReply listBBReply = new ListBBReply(sessionToken, bb.ListBillboards(listBBRequest.getSessionToken()));
+                    ListBBReply listBBReply = new ListBBReply(sessionToken, bb.ListBillboards());
                     oos.writeObject(listBBReply);
                 } catch (Exception e) {
                     generalReply = new GeneralReply(sessionToken,false);
@@ -661,14 +713,44 @@ public class Server {
             }
             oos.flush();
         } else if (clientRequest instanceof ScheduleBillboardRequest) {
-            ScheduleBillboardRequest temp = (ScheduleBillboardRequest) clientRequest;
-            ScheduleSQL Schedule = new ScheduleSQL();
-            Schedule.ScheduleBillboard(temp.getBillboardName(), temp.getScheduledTime(), temp.getDuration(), temp.getReoccurType(), temp.getReoccurAmount());
+            ScheduleBillboardRequest scheduleBillboardRequest = (ScheduleBillboardRequest) clientRequest;
+            SessionToken sessionToken = findSessionToken(scheduleBillboardRequest.getSessionToken());
+            if (!tokenCheck(scheduleBillboardRequest.getSessionToken())) {
+                sessionTokens.remove(sessionToken);
+                LogoutReply logoutReply = new LogoutReply(true);
+                oos.writeObject(logoutReply);
+            } else {
+                // Reset the used time of the session token.
+                sessionToken.setUsedTime(LocalDateTime.now());
+                GeneralReply generalReply;
+                try {
+                    ScheduleSQL Schedule = new ScheduleSQL();
+                    Schedule.ScheduleBillboard(scheduleBillboardRequest.getBillboardName(), scheduleBillboardRequest.getScheduledTime(), scheduleBillboardRequest.getDuration(), scheduleBillboardRequest.getReoccurType(), scheduleBillboardRequest.getReoccurAmount());
+                } catch (Exception e) {
+                    generalReply = new GeneralReply(sessionToken, false);
+                    oos.writeObject(generalReply);
+                }
+            }
             oos.flush();
         } else if (clientRequest instanceof DeleteScheduleRequest) {
-            DeleteScheduleRequest temp = (DeleteScheduleRequest) clientRequest;
-            ScheduleSQL Schedule = new ScheduleSQL();
-            Schedule.DeleteSchedule(temp.getScheduledName(), temp.getScheduledTime());
+            DeleteScheduleRequest deleteScheduleRequest = (DeleteScheduleRequest) clientRequest;
+            SessionToken sessionToken = findSessionToken(deleteScheduleRequest.getSessionToken());
+            if (!tokenCheck(deleteScheduleRequest.getSessionToken())) {
+                sessionTokens.remove(sessionToken);
+                LogoutReply logoutReply = new LogoutReply(true);
+                oos.writeObject(logoutReply);
+            } else {
+                // Reset the used time of the session token.
+                sessionToken.setUsedTime(LocalDateTime.now());
+                GeneralReply generalReply;
+                try {
+                    ScheduleSQL Schedule = new ScheduleSQL();
+                    Schedule.DeleteSchedule(deleteScheduleRequest.getScheduledName(), deleteScheduleRequest.getScheduledTime());
+                } catch (Exception e) {
+                    generalReply = new GeneralReply(sessionToken, false);
+                    oos.writeObject(generalReply);
+                }
+            }
             oos.flush();
         } else if (clientRequest instanceof WeeklyScheduleRequest) {
             WeeklyScheduleRequest weeklyscheduleRequest = (WeeklyScheduleRequest) clientRequest;
@@ -676,7 +758,40 @@ public class Server {
             WeeklyScheduleReply weeklyscheduleReply = new WeeklyScheduleReply(Schedule.ScheduledInformation());
             oos.writeObject(weeklyscheduleReply);
         }
-        // If the request in an instance of import/export billboard request
+        else if (clientRequest instanceof GetCurrentScheduledRequest) {
+            GetCurrentScheduledRequest GetCurrentScheduledRequest = (GetCurrentScheduledRequest) clientRequest;
+            ScheduleSQL Schedule = new ScheduleSQL();
+            GetCurrentScheduledReply getcurrentscheduledReply = new GetCurrentScheduledReply(Schedule.GetTitleCurrentScheduled(), true);
+            oos.writeObject(getcurrentscheduledReply);
+        }
+        else if (clientRequest instanceof ListScheduleRequest) {
+            ListScheduleRequest listScheduleRequest = (ListScheduleRequest) clientRequest;
+            SessionToken sessionToken = findSessionToken(listScheduleRequest.getSessionToken());
+
+            // Remove session token from the list and send a logout request if it expired.
+            if (!tokenCheck(listScheduleRequest.getSessionToken())) {
+                sessionTokens.remove(sessionToken);
+                LogoutReply logoutReply = new LogoutReply(true);
+                oos.writeObject(logoutReply);
+            } else {
+                // Reset the used time of the session token.
+                sessionToken.setUsedTime(LocalDateTime.now());
+                GeneralReply generalReply;
+
+                // Return a reply object to the client containing JTable of billboards
+                try {
+                    ScheduleSQL Schedule = new ScheduleSQL();
+                    ListScheduleReply listScheduleReply = new ListScheduleReply(sessionToken, Schedule.GetScheduledInfo(sessionToken));
+                    oos.writeObject(listScheduleReply);
+                } catch (Exception e) {
+                    generalReply = new GeneralReply(sessionToken, false);
+                    oos.writeObject(generalReply);
+                }
+
+            }
+            oos.flush();
+        }
+            // If the request in an instance of import/export billboard request
         else if (clientRequest instanceof XmlRequest) {
             XmlRequest xmlRequest = (XmlRequest) clientRequest;
             SessionToken sessionToken = findSessionToken(xmlRequest.getSessionToken());
@@ -694,8 +809,17 @@ public class Server {
                 // Return a reply object to the client containing JTable of billboards
                 try {
                     // if the user exports a billboard
-                    if (xmlRequest.getXmlFile() == null) {
+                    if (xmlRequest.getXmlFile() == null && xmlRequest.isExportState()) {
+                        String xmlName = xmlRequest.getXmlName() + ".xml";
+                        File exportFile = new File("src/xmlBillboards/" + xmlName);
+                        XmlReply xmlReply = new XmlReply(sessionToken, exportFile);
+                        oos.writeObject(xmlReply);
+                    }
+                    // if the user creates a billboard
+                    if (xmlRequest.getXmlFile() == null && !xmlRequest.isExportState()) {
                         new MakeXMLFile(xmlRequest.getXmlName());
+                        generalReply = new GeneralReply(sessionToken, true);
+                        oos.writeObject(generalReply);
                     }
                     // if the user imports a billboard
                     else {
@@ -709,16 +833,16 @@ public class Server {
                         BillboardSQL bb = new BillboardSQL();
                         bb.CreateBillboard(billboardName, xmlRequest.getUserName(), extractFromXML.TxtColourStr, extractFromXML.BGColourStr,
                                 extractFromXML.message, extractFromXML.image, extractFromXML.information, extractFromXML.InfoColourStr);
+                        generalReply = new GeneralReply(sessionToken, true);
+                        oos.writeObject(generalReply);
                     }
-                    generalReply = new GeneralReply(sessionToken, true);
                 } catch (Exception e) {
                     generalReply = new GeneralReply(sessionToken,false);
+                    oos.writeObject(generalReply);
                 }
-                oos.writeObject(generalReply);
                 oos.flush();
             }
         }
     }
 }
-
